Create GET and POST request handlers for books (bookings).
Create new file in views, place_book.py, to house instructions for
/places/<place_id>/books route.

Create GET and POST handlers.
-format string value from request.form into a date time object for
the date_start attribute, using strptime. (This is necessary to
conform to specs of PlaceBook model).

Change values for place_id and user_id in PlaceBook's to_hash method.
(Add .id, so that hash receives the int values for the ids of those
records instead of Peewee objects.)