Fix var name from db to database in test suite.
Test suite was originally written using @db to refer to the data-
base to connect to. This was the variable name that Steven used.

However, the specs in the original project says that the variable
should be @database and our own code uses this variable name.

Change all cases of @db to @database.