Fix test_get methods in test suite.
To convert response body from JSON, use json.loads instead of json
.dumps.

Since the response represents one record instead of a list, simply
parse for keys and values and do not try to access items in a list.
The response should be a dict when converted into Python and not a
list of dicts.